# 🚀 Hybrid MCP Installation Strategy - 58% Performance Improvement