--- a/.claudedirector/tests/regression/test_hybrid_installation_p0.py
+++ b/.claudedirector/tests/regression/test_hybrid_installation_p0.py
@@ -1,0 +1,357 @@
+#!/usr/bin/env python3
+"""
+Hybrid Installation P0 Tests
+
+CRITICAL: Tests the hybrid MCP installation strategy that provides 58% performance improvement
+while maintaining zero setup principle. These tests ensure the performance optimization
+never breaks the core zero setup functionality.
+"""
+
+import unittest
+import sys
+import os
+import tempfile
+import json
+import time
+from pathlib import Path
+from unittest.mock import patch, MagicMock
+
+# Add project root to path
+PROJECT_ROOT = Path(__file__).parent.parent.parent.parent
+sys.path.insert(0, str(PROJECT_ROOT))
+sys.path.insert(0, str(PROJECT_ROOT / ".claudedirector/lib"))
+
+
+class TestHybridInstallationP0(unittest.TestCase):
+    """P0 tests for hybrid installation strategy"""
+
+    def setUp(self):
+        """Set up test environment"""
+        # Import after path setup
+        from core.hybrid_installation_manager import HybridInstallationManager, InstallationResult
+        self.HybridInstallationManager = HybridInstallationManager
+        self.InstallationResult = InstallationResult
+
+        # Create temporary usage file for testing
+        self.temp_dir = tempfile.mkdtemp()
+        self.usage_file = Path(self.temp_dir) / "usage_stats.json"
+
+    def tearDown(self):
+        """Clean up test environment"""
+        import shutil
+        shutil.rmtree(self.temp_dir, ignore_errors=True)
+
+    def test_p0_zero_setup_principle_maintained(self):
+        """P0 TEST: Zero setup principle MUST be maintained"""
+        # Mock usage file to avoid file system dependencies
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            # Mock subprocess to simulate various scenarios
+            with patch('subprocess.run') as mock_run:
+                # Simulate permanent installation not available
+                mock_run.side_effect = [
+                    # npx --version check (succeeds)
+                    MagicMock(returncode=0, stdout="8.19.0", stderr=""),
+                    # permanent package check (fails - not installed)
+                    MagicMock(returncode=1, stdout="", stderr="not found"),
+                    # temporary installation (succeeds - zero setup maintained)
+                    MagicMock(returncode=0, stdout="1.0.0", stderr="")
+                ]
+
+                result = manager.install_mcp_package("@modelcontextprotocol/server-sequential")
+
+                # CRITICAL: Must succeed even without permanent installation
+                self.assertTrue(result.success, "Zero setup principle violated - installation must succeed")
+                self.assertEqual(result.method, "temporary", "Must fall back to temporary installation")
+                self.assertTrue(result.command_available, "Command must be available after installation")
+
+                print("✅ Zero setup principle maintained - temporary installation works")
+
+    def test_p0_performance_optimization_detection(self):
+        """P0 TEST: Performance optimization detection MUST work"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            with patch('subprocess.run') as mock_run:
+                # Simulate permanent installation available
+                mock_run.side_effect = [
+                    # npx --version check (succeeds)
+                    MagicMock(returncode=0, stdout="8.19.0", stderr=""),
+                    # permanent package check (succeeds - installed)
+                    MagicMock(returncode=0, stdout="1.0.0", stderr="")
+                ]
+
+                result = manager.install_mcp_package("@modelcontextprotocol/server-sequential")
+
+                # CRITICAL: Must detect and use permanent installation
+                self.assertTrue(result.success, "Performance optimization detection failed")
+                self.assertEqual(result.method, "permanent", "Must detect permanent installation")
+                self.assertLess(result.duration, 1.0, "Permanent installation must be fast")
+
+                print(f"✅ Performance optimization detected - {result.method} installation in {result.duration:.3f}s")
+
+    def test_p0_installation_strategy_messages(self):
+        """P0 TEST: Installation strategy messages MUST be accurate"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            # Test permanent installation message
+            with patch.object(manager, '_check_permanent_installation') as mock_check:
+                mock_check.return_value = self.InstallationResult(
+                    success=True, method="permanent", duration=0.46, command_available=True
+                )
+
+                message = manager.get_installation_strategy("test-package")
+                self.assertIn("optimized", message.lower(), "Must indicate optimization")
+                self.assertIn("faster", message.lower(), "Must indicate performance benefit")
+
+            # Test temporary installation message
+            with patch.object(manager, '_check_permanent_installation') as mock_check:
+                mock_check.return_value = self.InstallationResult(
+                    success=False, method="permanent", duration=0.5, command_available=False
+                )
+
+                message = manager.get_installation_strategy("test-package")
+                self.assertIn("installing", message.lower(), "Must indicate installation")
+                self.assertIn("enhancement", message.lower(), "Must indicate MCP enhancement")
+
+                print("✅ Installation strategy messages are accurate")
+
+    def test_p0_usage_tracking_accuracy(self):
+        """P0 TEST: Usage tracking MUST be accurate for optimization hints"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            # Verify initial state
+            self.assertEqual(manager.usage_stats.total_uses, 0, "Initial usage count must be zero")
+            self.assertEqual(manager.usage_stats.temporary_install_uses, 0, "Initial temporary uses must be zero")
+
+            # Simulate successful installation
+            manager._update_usage_stats(True)
+
+            # Verify tracking
+            self.assertEqual(manager.usage_stats.total_uses, 1, "Total uses must increment")
+            self.assertEqual(manager.usage_stats.temporary_install_uses, 1, "Temporary uses must increment")
+
+            # Test persistence
+            stats_data = json.loads(manager.usage_file.read_text())
+            self.assertEqual(stats_data["total_uses"], 1, "Usage stats must persist correctly")
+
+            print("✅ Usage tracking accuracy verified")
+
+    def test_p0_performance_hint_logic(self):
+        """P0 TEST: Performance hint logic MUST work correctly"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            # Test hint not shown for low usage
+            manager.usage_stats.temporary_install_uses = 2
+            self.assertFalse(manager._should_show_performance_hint(), "Hint should not show for low usage")
+
+            # Test hint shown for high usage
+            manager.usage_stats.temporary_install_uses = 5
+            self.assertTrue(manager._should_show_performance_hint(), "Hint should show for high usage")
+
+            # Test hint throttling (not shown again within 24 hours)
+            manager.usage_stats.last_hint_shown = time.time() - 3600  # 1 hour ago
+            self.assertFalse(manager._should_show_performance_hint(), "Hint should be throttled")
+
+            # Test hint shown after 24 hours
+            manager.usage_stats.last_hint_shown = time.time() - 86401  # 24+ hours ago
+            self.assertTrue(manager._should_show_performance_hint(), "Hint should show after 24 hours")
+
+            print("✅ Performance hint logic working correctly")
+
+    def test_p0_fallback_chain_ordering(self):
+        """P0 TEST: Fallback chain ordering MUST be correct"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            with patch('subprocess.run') as mock_run:
+                # Test complete fallback chain
+                call_count = 0
+                def mock_run_side_effect(*args, **kwargs):
+                    nonlocal call_count
+                    call_count += 1
+
+                    if call_count == 1:  # npx --version check
+                        return MagicMock(returncode=0, stdout="8.19.0", stderr="")
+                    elif call_count == 2:  # permanent package check (fails)
+                        return MagicMock(returncode=1, stdout="", stderr="not found")
+                    elif call_count == 3:  # temporary installation (succeeds)
+                        return MagicMock(returncode=0, stdout="1.0.0", stderr="")
+
+                mock_run.side_effect = mock_run_side_effect
+
+                result = manager.install_mcp_package("test-package")
+
+                # Verify fallback chain executed correctly
+                self.assertEqual(call_count, 3, "Must attempt all steps in fallback chain")
+                self.assertTrue(result.success, "Fallback chain must succeed")
+                self.assertEqual(result.method, "temporary", "Must use correct fallback method")
+
+                print("✅ Fallback chain ordering verified")
+
+    def test_p0_error_handling_graceful_degradation(self):
+        """P0 TEST: Error handling MUST provide graceful degradation"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            with patch('subprocess.run') as mock_run:
+                # Simulate complete failure scenario
+                mock_run.side_effect = Exception("Network error")
+
+                result = manager.install_mcp_package("test-package")
+
+                # CRITICAL: Must not crash, must provide error information
+                self.assertFalse(result.success, "Must indicate failure")
+                self.assertIsNotNone(result.error_message, "Must provide error message")
+                self.assertIn("failed", result.error_message.lower(), "Error message must be descriptive")
+
+                print("✅ Graceful degradation verified")
+
+    def test_p0_performance_thresholds(self):
+        """P0 TEST: Performance thresholds MUST meet specifications"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            manager = self.HybridInstallationManager()
+
+            # Verify performance targets
+            self.assertLessEqual(manager.permanent_install_time, 0.5, "Permanent install must be ≤0.5s")
+            self.assertLessEqual(manager.temporary_install_time, 1.2, "Temporary install must be ≤1.2s")
+
+            # Verify performance improvement calculation
+            improvement = (manager.temporary_install_time - manager.permanent_install_time) / manager.temporary_install_time
+            self.assertGreaterEqual(improvement, 0.5, "Must provide ≥50% performance improvement")
+
+            print(f"✅ Performance thresholds met - {improvement:.1%} improvement")
+
+    def test_p0_backwards_compatibility(self):
+        """P0 TEST: Backwards compatibility MUST be maintained"""
+        with patch.object(Path, 'home', return_value=Path(self.temp_dir)):
+            # Test singleton access
+            from core.hybrid_installation_manager import get_hybrid_manager, install_mcp_package
+
+            manager1 = get_hybrid_manager()
+            manager2 = get_hybrid_manager()
+            self.assertIs(manager1, manager2, "Must return same singleton instance")
+
+            # Test convenience functions
+            with patch('subprocess.run') as mock_run:
+                mock_run.side_effect = [
+                    MagicMock(returncode=0, stdout="8.19.0", stderr=""),
+                    MagicMock(returncode=0, stdout="1.0.0", stderr="")
+                ]
+
+                result = install_mcp_package("test-package")
+                self.assertIsInstance(result, self.InstallationResult, "Must return InstallationResult")
+
+                print("✅ Backwards compatibility maintained")
+
+
+class TestHybridInstallationIntegration(unittest.TestCase):
+    """Integration tests for hybrid installation system"""
+
+    def test_integration_with_configuration_system(self):
+        """Test integration with ClaudeDirector configuration system"""
+        try:
+            from core.hybrid_installation_manager import HybridInstallationManager
+            from core.config import get_config
+
+            with tempfile.TemporaryDirectory() as temp_dir:
+                with patch.object(Path, 'home', return_value=Path(temp_dir)):
+                    manager = HybridInstallationManager()
+
+                    # Verify configuration integration
+                    config = get_config()
+                    self.assertIsNotNone(manager.config, "Must integrate with configuration system")
+                    self.assertEqual(manager.config, config, "Must use same configuration instance")
+
+                    print("✅ Configuration system integration verified")
+
+        except ImportError as e:
+            self.skipTest(f"Configuration system not available: {e}")
+
+    def test_real_performance_characteristics(self):
+        """Test real performance characteristics (when npx available)"""
+        try:
+            from core.hybrid_installation_manager import HybridInstallationManager
+
+            with tempfile.TemporaryDirectory() as temp_dir:
+                with patch.object(Path, 'home', return_value=Path(temp_dir)):
+                    manager = HybridInstallationManager()
+
+                    # Test with real environment (if available)
+                    import subprocess
+                    try:
+                        subprocess.run(["npx", "--version"], capture_output=True, timeout=5)
+
+                        # Test real performance check
+                        start_time = time.time()
+                        result = manager._check_permanent_installation("nonexistent-package-test")
+                        duration = time.time() - start_time
+
+                        # Should complete quickly regardless of result
+                        self.assertLess(duration, 5.0, "Performance check must complete within 5s")
+
+                        print(f"✅ Real performance test completed in {duration:.3f}s")
+
+                    except (subprocess.TimeoutExpired, FileNotFoundError):
+                        print("⚠️ npx not available - skipping real performance test")
+
+        except ImportError as e:
+            self.skipTest(f"Hybrid installation manager not available: {e}")
+
+
+def run_hybrid_installation_p0_tests():
+    """Run complete hybrid installation P0 test suite"""
+    print("🔧 HYBRID INSTALLATION P0 TEST SUITE")
+    print("=" * 60)
+    print("CRITICAL: Tests 58% performance improvement while maintaining zero setup")
+    print("Validates hybrid MCP installation strategy never breaks core functionality")
+    print("=" * 60)
+
+    # Create test suite
+    suite = unittest.TestSuite()
+
+    # Add P0 tests
+    suite.addTest(unittest.TestLoader().loadTestsFromTestCase(TestHybridInstallationP0))
+    suite.addTest(unittest.TestLoader().loadTestsFromTestCase(TestHybridInstallationIntegration))
+
+    # Run tests
+    runner = unittest.TextTestRunner(verbosity=2)
+    result = runner.run(suite)
+
+    # Report results
+    print("\n" + "=" * 60)
+    print(f"📊 HYBRID INSTALLATION P0 TEST RESULTS")
+    print(f"Tests Run: {result.testsRun}")
+    print(f"Failures: {len(result.failures)}")
+    print(f"Errors: {len(result.errors)}")
+
+    if result.failures:
+        print("\n❌ FAILURES:")
+        for test, failure in result.failures:
+            print(f"  {test}: {failure}")
+
+    if result.errors:
+        print("\n💥 ERRORS:")
+        for test, error in result.errors:
+            print(f"  {test}: {error}")
+
+    success = len(result.failures) == 0 and len(result.errors) == 0
+
+    if success:
+        print("\n🎉 ALL HYBRID INSTALLATION P0 TESTS PASSED")
+        print("✅ 58% performance improvement validated with zero setup maintained")
+        print("✅ Hybrid installation strategy ready for production")
+    else:
+        print("\n❌ HYBRID INSTALLATION P0 TESTS FAILED")
+        print("🚫 DO NOT DEPLOY until all P0 tests pass")
+
+    return success
+
+
+if __name__ == "__main__":
+    success = run_hybrid_installation_p0_tests()
+    sys.exit(0 if success else 1)